+++ app-client/src/containers/user/Dashboard.tsx
@@ -10,7 +10,7 @@ import * as React from 'react';
 import * as Radium from 'radium';
 import {RouteComponentProps} from 'react-router';
 import {Container} from 'src/components/ReusableComponents';
-import {NavbarHeader} from 'src/components/NavbarHeader/NavbarHeader';
+import {NavbarHeader} from 'src/components/navbarHeader/NavbarHeader';
 
 @Radium
 export class Dashboard extends React.Component<RouteComponentProps<void>> {