+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 
 - Added Cognito SDKs.
 - Added Login and Signup page.
+- Added Forgot password and Reset password page.